Restructuracion de la carpeta components
