+++ src/components/cards/CardDesignerUX.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .card-designerux{
     width: 300px;
