+++ src/components/fig/LineColumn.jsx
