+++ src/components/fig/LineColumn.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .linecolumn {
     border-left: 3px solid $grey;
