+++ src/components/fig/LineRow.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .box-linerow{
     display: flex;
