+++ src/components/header/LinkNavbar.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .linknav {
     color: $white;
