+++ src/components/pages/LinkSocialDesk.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 $changeBackground: (
     face: $colorFacebook,
