+++ src/components/pages/SocialDesk.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .box-socialdesk{
     position: fixed;
