+++ src/components/pages/TitleWeb.scss
@@ -1,4 +1,4 @@
-@import "../../../sass/Style.scss";
+@import "../../sass/Style.scss";
 
 .box-titleweb{
     display: flex;
