+++ src/components/pages/home/Designer.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/Style.scss";
+@import "../../../sass/Style.scss";
 
 .box-designer{
     display: flex;
