+++ src/components/pages/home/EpigraphName.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/Style.scss";
+@import "../../../sass/Style.scss";
 
 .my-name{
     display: flex;
