+++ src/components/pages/home/SubtitleHero.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/Style.scss";
+@import "../../../sass/Style.scss";
 
 .subtitle{
     display: flex;
